// src/app/compartilhado/models/prestador.model.ts
export class Prestador {
    id: number;
    telefone: string;
    celular: string;
    cep: string;
    bairro: string;
    cidade: string;
    estado: string;
    numero: string;
    descricao: string;
    avaliacao: string;
    foto: string;
    tipo: string;
    usuario_id: number;
    id_serv_1: number;
    id_serv_2: number;
    id_serv_3: number;
}